requirements.txt
streamlit==1.37.1
pypdf==4.3.1
numpy==2.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1